more refactoring, have to think more about what methods i should actually move to parent class
